Change argument list of GETGROUPS/SETGROUPS to represent the argument list of the syscalls. getgroups() first argument is a signed int!. Signed-off-by: Daniel Gollub <dgollub@suse.de>. Also, setgroups02(_16) is working with an uninitialized list, expect the first GID field. setgroups() and getgroups() additionally get called with the full length  of the uninitialized list. The uninitialized values cause setgroups02_16 to fail: === setgroups02_16    1  BROK  :  gid returned from getgroups is too large for testing setgroups32 === Since the uninitialized values could be greater then 0xffff. Signed-off-by: Daniel Gollub <dgollub@suse.de>.

+++ testcases/kernel/syscalls/setgroups/compat_16.h
@@ -32,13 +32,13 @@ extern void cleanup(void);
 #ifdef TST_USE_COMPAT16_SYSCALL
 
 long
-SETGROUPS(int gidsetsize, GID_T *list)
+SETGROUPS(size_t gidsetsize, GID_T *list)
 {
 	return syscall(__NR_setgroups, gidsetsize, list);
 }
 
 int
-GETGROUPS(size_t size16, GID_T *list16)
+GETGROUPS(int size16, GID_T *list16)
 {
 	int r;
 	int i;
@@ -75,7 +75,7 @@ SETGROUPS(size_t size, const GID_T *list)
 }
 
 int
-GETGROUPS(size_t size, GID_T *list)
+GETGROUPS(int size, GID_T *list)
 {
 	return getgroups(size, list);
 }
